pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "warpfield"
version = "0.0.1"
description = "GPU-accelerated 3D non-rigid registration"
authors = [
    { name = "jlab.berlin" },
    { name = "Maximilian Hoffmann" },
    { name = "Benjamin Judkewitz" },
]
license = { text = "MIT" }
readme = "README.md"
keywords = ["warpfield", "python"]
requires-python = ">=3.9"
dependencies = [
    "numpy<2",
    "scipy",
    "cupy<13.3",
    "tqdm",
    "pydantic",
    "pyyaml",
    "h5py",
    "hdf5plugin",
    "nibabel",
    "pydicom",
    "tifffile",
]

[project.urls]
Homepage = "https://github.com/danionella/warpfield"

[tool.setuptools.packages.find]
where = ["src"]
include = ["warpfield*"]

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]

[tool.black]
line-length = 120
skip_magic_trailing_comma = true
